requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.13
django-allauth==0.51.0
django-environ==0.8.1
django-paypal==2.0
django-phonenumber-field==7.0.2
django-sslserver==0.22
django-storages==1.12.3
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.1
oauthlib==3.2.0
phonenumbers==8.13.7
Pillow==9.4.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2021.3
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
stripe==5.2.0
toml==0.10.2
typing_extensions==4.5.0
urllib3==1.26.11
whitenoise==5.3.0
